[Pytorch Edge] Refactor Compatibility Stuff into own directory
Differential Revision: [D33640527](https://our.internmc.facebook.com/intern/diff/D33640527/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D33640527/)!

[ghstack-poisoned]
