+++ torch/csrc/jit/mobile/compatibility/runtime_compatibility.cpp
@@ -1,7 +1,7 @@
 #include <ATen/core/dispatch/Dispatcher.h>
 #include <ATen/core/type_factory.h>
 #include <caffe2/serialize/inline_container.h>
-#include <torch/csrc/jit/mobile/runtime_compatibility.h>
+#include <torch/csrc/jit/mobile/compatibility/runtime_compatibility.h>
 #include <torch/csrc/jit/mobile/type_parser.h>
 #include <torch/csrc/jit/runtime/operator.h>
 #include <torch/custom_class.h>
